requirements.txt
# Google API Dependencies
google-api-python-client>=2.86.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
google-auth>=2.0.0

# Authentication & Security
pyjwt>=2.0.0
cryptography>=40.0.0

# Web API Dependencies
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0.3
requests>=2.25.0
itsdangerous>=2.0.0
starlette>=0.27.0

# Database Dependencies
sqlalchemy>=1.4.0

# AI & LLM Dependencies
ollama>=0.1.5
langchain>=0.1.0
langchain-core>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2 # For formatting standards
langchain-experimental>=0.0.36

# Utility Dependencies
python-dotenv>=1.0.0
click>=8.1.3
rich>=13.4.2
pydantic>=2.0.0
pydantic-settings>=2.0.0  # Required for BaseSettings
python-dateutil>=2.8.2
dateparser>=1.1.8
asyncio>=3.4.3
